extract ImageDataWrapper to zarrimagedatawrapper.h
